+++ components/grid-elements.vue
@@ -1,83 +1,12 @@
 <script setup lang="ts">
-
-import interact from 'interactjs'
-import { useWindowSize } from '@vueuse/core'
-
 // TODO temp
 const link = "https://www.calculatorsoup.com/calculators/math/percentage.php"
-
 const editModel = defineModel('edit', { default: false });
-
-const { width, height } = useWindowSize();
-const yGrid = ref(height.value / 10);
-const xGrid = ref(width.value / 10);
-const yMin = ref(yGrid.value);
-const xMin = ref(xGrid.value);
-
 const gridElements = ref<HTMLDivElement | null>(null);
 
 onMounted(() => {
-    const initializeGridElements = (element: HTMLElement) => {
-        let y = 0;
-        let x = 0;
-        interact(element).draggable({
-            listeners: {
-                move(event) {
-                    x += event.dx;
-                    y += event.dy;
 
-                    // expand container when needed
-                    const containerRect = gridElements.value!.getBoundingClientRect();
-                    const relativeTop = event.rect.top - containerRect.top;
-                    const actualTop = event.rect.height + relativeTop;
-                    if (actualTop > containerRect.height - yGrid.value) {
-                        gridElements.value!.style.height = `${containerRect.height + window.innerHeight}px`;
-                    }
-
-                    event.target.style.transform = `translate(${x}px,${y}px)`;
-                },
-            },
-            modifiers: [
-                interact.modifiers.snap({
-                    targets: [interact.snappers.grid({ x: xGrid.value, y: yGrid.value })],
-                    relativePoints: [{ x: 0, y: 0 },]
-                }),
-                interact.modifiers.restrict({
-                    restriction: 'parent',
-                    elementRect: { top: 0, left: 0, bottom: 1, right: 1 },
-                })
-            ],
-        }).resizable({
-            edges: { left: true, right: true, bottom: true, top: true },
-            listeners: {
-                move(event) {
-                    const target = event.target
-                    target.style.width = `${event.rect.width}px`
-                    target.style.height = `${event.rect.height}px`
-                }
-            },
-            modifiers: [
-                interact.modifiers.restrictEdges({
-                    outer: 'parent'
-                }),
-                interact.modifiers.restrictSize({
-                    min: { width: xMin.value, height: yMin.value }
-                }),
-                interact.modifiers.snap({
-                    targets: [interact.snappers.grid({ x: xGrid.value, y: yGrid.value })],
-                    relativePoints: [{ x: 0, y: 0 },]
-                }),
-            ],
-        })
-    }
-
-    gridElements.value!.style.height = `${gridElements.value!.offsetHeight}px`;
-    for (const element of gridElements.value!.children as unknown as HTMLElement[]) {
-        element.style.height = `${yGrid.value}px`;
-        element.style.width = `${xGrid.value}px`;
-        initializeGridElements(element);
-    }
-})
+});
 </script>
 
 <template>
@@ -98,7 +27,9 @@ onMounted(() => {
 .grid-elements {
     width: 100%;
     min-height: 100dvh;
-    user-select: none;
+    display: grid;
+    grid-template-columns: repeat(10, 1fr);
+    grid-template-rows: repeat(auto-fill, 10vh);
 
     .iframe {
         border: unset;
